+++ src/etools/applications/partners/permission_matrix/intervention_permissions.csv
@@ -4,7 +4,7 @@ Field no,Field Name,Group,Condition,Status,Action,Allowed
 3.2.4,activation_letter_attachment,*,contingency on,Active,Required,TRUE
 3.2.4,agreement,Partnership Manager,unlocked,Draft,Edit,TRUE
 3.2.5,agreement,*,,*,Required,TRUE
-3.2.7,amendments,*,not_in_amendment_mode,*,View,TRUE
+3.2.7,amendments,UNICEF User,not_in_amendment_mode,*,View,TRUE
 3.2.8,amendments,Partnership Manager,not_in_amendment_mode,Signed,Edit,TRUE
 3.2.9,amendments,Partnership Manager,not_in_amendment_mode,Active,Edit,TRUE
 3.3.7,attachments,*,not_in_amendment_mode,*,View,TRUE
