+++ src/etools/applications/partners/tests/test_api_interventions.py
@@ -3,27 +3,24 @@ import json
 from unittest import skip
 
 from django.core.cache import cache
-from django.core.files.uploadedfile import SimpleUploadedFile
 from django.db import connection
 from django.test import SimpleTestCase
 from django.urls import resolve, reverse
 from django.utils import timezone
 
 from rest_framework import status
-from rest_framework.exceptions import ErrorDetail
 from rest_framework.test import APIRequestFactory
 from unicef_attachments.models import Attachment
 from unicef_locations.tests.factories import LocationFactory
 from unicef_snapshot.models import Activity
 
-from etools.applications.attachments.models import AttachmentFlat
-from etools.applications.attachments.tests.factories import AttachmentFactory, AttachmentFileTypeFactory
+from etools.applications.attachments.tests.factories import AttachmentFactory
 from etools.applications.core.tests.cases import BaseTenantTestCase
 from etools.applications.core.tests.mixins import URLAssertionMixin
 from etools.applications.environment.helpers import tenant_switch_is_active
 from etools.applications.environment.tests.factories import TenantSwitchFactory
 from etools.applications.funds.tests.factories import FundsReservationItemFactory
-from etools.applications.partners.models import Intervention, InterventionAmendment, InterventionResultLink
+from etools.applications.partners.models import Intervention, InterventionResultLink
 from etools.applications.partners.permissions import InterventionPermissions
 from etools.applications.partners.tests.factories import (
     AgreementFactory,
@@ -97,7 +94,6 @@ class TestInterventionsAPI(BaseTenantTestCase):
             "agreement",
             "agreement_id",
             "amendment",
-            "amendments",
             "attachments",
             "budget_owner",
             "budget_owner_id",
@@ -1741,315 +1737,6 @@ class TestInterventionAmendmentListAPIView(BaseTenantTestCase):
         response_json = json.loads(response.rendered_content)
         self.assertIsInstance(response_json, list)
         self.assertFalse(response_json)
-
-
-class TestInterventionAmendmentCreateAPIView(BaseTenantTestCase):
-    def setUp(self):
-        super().setUp()
-
-        self.partnership_manager_user = UserFactory(is_staff=True, groups__data=['Partnership Manager', 'UNICEF User'])
-
-        self.intervention = InterventionFactory(status=Intervention.SIGNED)
-        self.url = reverse(
-            "partners_api:intervention-amendments-add",
-            kwargs={'intervention_pk': self.intervention.id}
-        )
-
-        self.uploaded_file = SimpleUploadedFile('hello_world.txt', 'hello world!'.encode('utf-8'))
-        self.data = {
-            "types": InterventionAmendment.DATES,
-            "signed_date": datetime.date.today(),
-            "signed_amendment": self.uploaded_file,
-        }
-        self.file_type_signed_amendment_attachment = AttachmentFileTypeFactory(
-            code="partners_intervention_amendment_signed"
-        )
-        self.file_type_internal_prc_review = AttachmentFileTypeFactory(
-            code="partners_intervention_amendment_internal_prc_review"
-        )
-
-    def assertResponseFundamentals(self, response):
-        """Assert common fundamentals about the response."""
-        response_json = json.loads(response.rendered_content)
-        self.assertIsInstance(response_json, dict)
-        self.assertIn('id', response_json)
-        return response_json
-
-    def test_no_permission_user_forbidden(self):
-        """Ensure a non-staff user gets the 403 smackdown"""
-        response = self._make_request(user=UserFactory())
-        self.assertEquals(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_unauthenticated_user_forbidden(self):
-        """Ensure an unauthenticated user gets the 403 smackdown"""
-        factory = APIRequestFactory()
-        view_info = resolve(self.url)
-        request = factory.post(self.url, data={})
-        response = view_info.func(request)
-        self.assertEquals(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_group_permission_partnership_member(self):
-        """Ensure group membership is sufficient for create;"""
-        user = UserFactory(is_staff=True)
-        response = self._make_request(user=user)
-        self.assertEquals(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_create_amendment_invalid_type(self):
-        invalid_type = 'invalid_choice'
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data={"types": [invalid_type], "signed_amendment": self.uploaded_file},
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEquals(response.data['types'],
-                          {0: [ErrorDetail(string='"invalid_choice" is not a valid choice.', code=f'{invalid_type}')]})
-
-    def test_create_amendment_other_type_no_description(self):
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data={
-                "types": [InterventionAmendment.OTHER],
-                "signed_amendment": self.uploaded_file,
-                "signed_date": datetime.date.today(),
-            },
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEquals(
-            response.data['non_field_errors'], [ErrorDetail(
-                string="Other description required, if type 'Other' selected.",
-                code='invalid'
-            )]
-        )
-
-    def test_create_amendment_invalid_file(self):
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data={},
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEquals(response.data['signed_amendment'], ['No file was submitted.'])
-
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data={'signed_amendment': 'asdf'},
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEquals(
-            response.data['signed_amendment'],
-            ['The submitted data was not a file. Check the encoding type on the form.']
-        )
-
-    def test_create_amendment_invalid_date(self):
-        tomorrow = datetime.date.today() + datetime.timedelta(days=1)
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data={
-                "signed_amendment": self.uploaded_file,
-                'signed_date': tomorrow,
-                'types': [InterventionAmendment.DATES, ],
-                'kind': InterventionAmendment.KIND_NORMAL,
-            },
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEquals(next(iter(response.data.values())), ['Date cannot be in the future!'])
-
-    def test_create_amendment_matching_date(self):
-        date = datetime.date.today() - datetime.timedelta(days=1)
-        InterventionAmendmentFactory(
-            intervention=self.intervention,
-            signed_date=date,
-        )
-        AttachmentFactory(
-            file="test_file.pdf",
-            file_type=None,
-            code="",
-        )
-        attachment_amendment = AttachmentFactory(
-            file="test_file_amendment.pdf",
-            file_type=None,
-            code="",
-        )
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data={
-                "signed_amendment_attachment": attachment_amendment.pk,
-                'signed_date': date,
-                'types': [InterventionAmendment.DATES, ],
-                'kind': InterventionAmendment.KIND_NORMAL,
-            },
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEquals(
-            next(iter(response.data.values())),
-            ['There is already an amendment with this signed date.'],
-        )
-
-    def test_create_amendment_success(self):
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data=self.data,
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_201_CREATED)
-        data = self.assertResponseFundamentals(response)
-        self.assertEquals(data['intervention'], self.intervention.id)
-
-    def test_create_amendment_with_signed_attachment(self):
-        attachment = AttachmentFactory(
-            file="test_file.pdf",
-            file_type=None,
-            code="",
-        )
-        self.data.pop("signed_amendment")
-        self.data["signed_amendment_attachment"] = attachment.pk
-        self.assertIsNone(attachment.file_type)
-        self.assertIsNone(attachment.content_object)
-        self.assertFalse(attachment.code)
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data=self.data,
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_201_CREATED)
-        data = self.assertResponseFundamentals(response)
-        self.assertEquals(data['intervention'], self.intervention.pk)
-        attachment_updated = Attachment.objects.get(pk=attachment.pk)
-        self.assertEqual(
-            attachment_updated.file_type.code,
-            self.file_type_signed_amendment_attachment.code
-        )
-        self.assertEqual(attachment_updated.object_id, data["id"])
-        self.assertEqual(
-            attachment_updated.code,
-            self.file_type_signed_amendment_attachment.code
-        )
-
-    def test_create_amendment_with_internal_prc_review(self):
-        attachment = AttachmentFactory(
-            file="test_file.pdf",
-            file_type=None,
-            code="",
-        )
-        flat_qs = AttachmentFlat.objects.filter(attachment=attachment)
-        assert flat_qs.exists()
-        flat = flat_qs.first()
-        assert not flat.partner
-        self.data["internal_prc_review"] = attachment.pk
-        self.assertIsNone(attachment.file_type)
-        self.assertIsNone(attachment.content_object)
-        self.assertFalse(attachment.code)
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data=self.data,
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_201_CREATED)
-        data = self.assertResponseFundamentals(response)
-        self.assertEquals(data['intervention'], self.intervention.pk)
-        attachment.refresh_from_db()
-        self.assertEqual(
-            attachment.file_type.code,
-            self.file_type_internal_prc_review.code
-        )
-        self.assertEqual(attachment.object_id, data["id"])
-        self.assertEqual(
-            attachment.code,
-            self.file_type_internal_prc_review.code
-        )
-
-        # check denormalization
-        flat = flat_qs.first()
-        assert flat.partner
-        assert flat.pd_ssfa
-        assert flat.pd_ssfa_number
-
-    def test_create_amendment_with_internal_prc_review_none(self):
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data=self.data,
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_201_CREATED)
-        data = self.assertResponseFundamentals(response)
-        self.assertEquals(data['intervention'], self.intervention.pk)
-
-    def test_create_amendment_when_already_in_amendment(self):
-        self.intervention.in_amendment = True
-        self.intervention.save()
-
-        response = self._make_request(
-            user=self.partnership_manager_user,
-            data=self.data,
-            request_format='multipart',
-        )
-
-        self.assertEquals(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEquals(
-            next(iter(response.data.values())),
-            ['Cannot add a new amendment while another amendment is in progress.']
-        )
-
-    def _make_request(self, user=None, data=None, request_format='json', **kwargs):
-        return self.forced_auth_req(
-            'post', self.url, user=user, data=data or {}, request_format=request_format, **kwargs
-        )
-
-
-class TestInterventionAmendmentDeleteView(BaseTenantTestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.unicef_staff = UserFactory(is_staff=True)
-        cls.intervention = InterventionFactory(status=Intervention.DRAFT)
-        cls.amendment = InterventionAmendmentFactory(
-            intervention=cls.intervention,
-            types=[InterventionAmendment.RESULTS],
-        )
-        cls.url = reverse(
-            "partners_api:intervention-amendments-del",
-            args=[cls.amendment.pk]
-        )
-
-    def test_delete(self):
-        response = self.forced_auth_req(
-            'delete',
-            self.url,
-            user=self.unicef_staff,
-        )
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
-
-    def test_delete_invalid(self):
-        self.intervention.status = Intervention.ACTIVE
-        self.intervention.save()
-        response = self.forced_auth_req(
-            'delete',
-            self.url,
-            user=self.unicef_staff,
-        )
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, ["You do not have permissions to delete an amendment"])
-
-    def test_intervention_amendments_delete(self):
-        response = self.forced_auth_req(
-            'delete',
-            reverse("partners_api:intervention-amendments-del", args=[404]),
-            user=self.unicef_staff,
-        )
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
 
 class TestInterventionListMapView(BaseTenantTestCase):
