+++ src/etools/applications/partners/tests/test_export_intervention.py
@@ -271,8 +271,8 @@ class TestInterventionAmendmentModelExport(BaseInterventionModelExportTestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         dataset = Dataset().load(response.content.decode('utf-8'), 'csv')
         self.assertEqual(dataset.height, 1)
-        self.assertEqual(len(dataset._get_headers()), 10)
-        self.assertEqual(len(dataset[0]), 10)
+        self.assertEqual(len(dataset._get_headers()), 15)
+        self.assertEqual(len(dataset[0]), 15)
 
     def test_csv_flat_export_api(self):
         response = self.forced_auth_req(
@@ -285,8 +285,8 @@ class TestInterventionAmendmentModelExport(BaseInterventionModelExportTestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         dataset = Dataset().load(response.content.decode('utf-8'), 'csv')
         self.assertEqual(dataset.height, 1)
-        self.assertEqual(len(dataset._get_headers()), 10)
-        self.assertEqual(len(dataset[0]), 10)
+        self.assertEqual(len(dataset._get_headers()), 16)
+        self.assertEqual(len(dataset[0]), 16)
 
 
 class TestInterventionResultModelExport(BaseInterventionModelExportTestCase):
